test.ino contains most of working code

+++ lighting_patterns/Final Code/clayPattern/Graph.hpp
@@ -57,8 +57,8 @@ public:
         return ret;
     };
     
-    uint8_t maxDist (int* dist) {
-        uint8_t Max = 0;
+    double maxDist (int* dist) {
+        double Max = 0;
         for(int i=0;i<_nV;i++) {
             if(dist[i] > Max) {
               Max = dist[i];
